+++ packages/ocap-kernel/src/index.test.ts
@@ -10,6 +10,7 @@ describe('index', () => {
       'ClusterConfigStruct',
       'Kernel',
       'KernelStatusStruct',
+      'RunLoopStatusStruct',
       'SubclusterStruct',
       'VatConfigStruct',
       'VatHandle',